+++ spec/lib/karafka/routing/topic_spec.rb
@@ -59,8 +59,8 @@ RSpec.describe_current do
 
   describe '#to_h' do
     let(:expected_keys) do
-      %i[kafka deserializer max_messages max_wait_time initial_offset] +
-        %i[id name consumer consumer_group_id subscription_group tags manual_offset_management]
+      %i[kafka deserializer max_messages max_wait_time initial_offset id name consumer] +
+        %i[consumer_group_id subscription_group manual_offset_management active_job]
     end
 
     it 'expect to contain all the topic attrs plus some inherited' do